+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -6,6 +6,7 @@
 - Update: Method chain statement builder can now omit semi-colons.
 - Update: Free Code Blocks now supported on Interfaces and Classes for cases where only freetext needs to be added a Class / Interface.
 - Fixed: Methods with abstract keyword and without statements will not output method body.
+- Fixed: ` : this()` constructor call didn't show when telling the Constructor to `CallsThis()` without arguments.
 
 ### Version 3.3.27